Fixed duplicat implementation in StorageEngineMock, made mockRegistry outlive transactionStatistics, added explicit, removed redundant includes
