+++ arangod/IResearch/IResearchDataStore.cpp
@@ -36,7 +36,6 @@
 #include "Metrics/ClusterMetricsFeature.h"
 #include "Metrics/Gauge.h"
 #include "Metrics/Guard.h"
-#include "Metrics/MetricsFeature.h"
 #include "RestServer/FlushFeature.h"
 #include "RestServer/DatabaseFeature.h"
 #include "RestServer/DatabasePathFeature.h"
